% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfactory.ffmq.R
\name{ffmq}
\alias{ffmq}
\title{Calculate FFMQ scores}
\usage{
ffmq(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
tibble
}
\description{
Calculates FFMQ scores.  Expects a data frame
formatted by \code{\link[expfactory:process_expfactory_survey]{expfactory::process_expfactory_survey}}, containing
responses from a single participant.
}
\keyword{FFMQ}
